crop-advisory-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.0
requests==2.31.0
openai==1.30.0
Pillow==10.3.0
